if we want to fork the JVM, don't even try to compile, it's not gonna work - instead, fork the JVM one level up, at the ScriptRunner

+++ core/src/main/scala/replpp/scripting/ScriptRunner.scala
@@ -6,8 +6,14 @@ import replpp.allPredefCode
 import scala.collection.immutable.{AbstractSeq, LinearSeq}
 import scala.jdk.CollectionConverters.*
 import scala.xml.NodeSeq
+import sys.process.Process
 
 object ScriptRunner {
+
+  def main(args: Array[String]): Unit = {
+    val config = Config.parse(args)
+    exec(config)
+  }
 
   def exec(config: Config): Unit = {
     val scriptFile = config.scriptFile.getOrElse(throw new AssertionError("scriptFile not defined"))
@@ -36,23 +42,39 @@ object ScriptRunner {
 //    val scriptContent = wrapForMainargs(predefCode, scriptCode)
     val scriptContent = wrapForMainargsFoo(predefCode, scriptCode)
     os.write.over(predefPlusScriptFileTmp, scriptContent)
-    val compilerArgs = replpp.compilerArgs(config, predefCode) :+ "-nowarn"
+    val verbose = replpp.verboseEnabled(config)
 
-    new ScriptingDriver(
-      compilerArgs = compilerArgs,
-      scriptFile = predefPlusScriptFileTmp.toIO,
-      scriptArgs = scriptArgs.toArray,
-      verbose = replpp.verboseEnabled(config)
-    ).compileAndRun() match {
-      case Some(exception) =>
-        System.err.println(s"error during script execution: ${exception.getMessage}")
-        System.err.println(s"note: line numbers may not be accurate - to help with debugging, the final scriptContent is at $predefPlusScriptFileTmp")
-        throw exception
-      case None => // no exception, i.e. all is good
-        System.err.println(s"script finished successfully")
-        // if the script failed, we don't want to delete the temporary file which includes the predef,
-        // so that the line numbers are accurate and the user can properly debug
-        os.remove(predefPlusScriptFileTmp)
+    val forkJvm = true // TODO get from config
+    // TODO refactor for readability
+    if (forkJvm) {
+      val args = Seq(
+        "-classpath",
+        replpp.classpath(config),
+        "replpp.scripting.ScriptRunner",
+        // TODO pass on config (unapplied), but remove `--fork`, so this doesn't become an endless loop
+        // TODO call ScriptRunner with non-fork", but pass on other config options -> unapply config params -> check scopt readme
+      )
+      if (verbose) println(s"forking jvm - executing `java ${args.mkString(" ")}`")
+      val p = Process("java", args).run()
+      assert(p.exitValue() == 0, s"error while invoking `java ${args.mkString(" ")}`. exit code was ${p.exitValue()}")
+    } else {
+      val compilerArgs = replpp.compilerArgs(config) :+ "-nowarn"
+      new ScriptingDriver(
+        compilerArgs = compilerArgs,
+        scriptFile = predefPlusScriptFileTmp.toIO,
+        scriptArgs = scriptArgs.toArray,
+        verbose = verbose
+      ).compileAndRun() match {
+        case Some(exception) =>
+          System.err.println(s"error during script execution: ${exception.getMessage}")
+          System.err.println(s"note: line numbers may not be accurate - to help with debugging, the final scriptContent is at $predefPlusScriptFileTmp")
+          throw exception
+        case None => // no exception, i.e. all is good
+          System.err.println(s"script finished successfully")
+          // if the script failed, we don't want to delete the temporary file which includes the predef,
+          // so that the line numbers are accurate and the user can properly debug
+          os.remove(predefPlusScriptFileTmp)
+      }
     }
   }
